Master function

This function does all of the steps in a single unified function.

corr_meta(
  data,
  n,
  type = c("average", "weighted", "simple"),
  missing = "remove",
  variable_names = NULL,
  ID = NULL,
  effect_size,
  var_cor,
  weights = NULL,
  moderators = NULL,
  random_params = NULL,
  structure = "CS",
  test = "t",
  intercept = FALSE,
  estimation_method = "REML",
  model = NULL,
  num_obs = NULL,
  ...
)

Arguments

data
Raw data input

n
Sample size

type
type of adjustment (olkin siotani)

missing
What to do with missing data

variable_names
An optional character vector specifying variable names of original data. This is mostly used to pass to df_to_corr. A character vector of variable names representing the columns to convert to a pairwise correlation matrix. Variable names take the following format: "columnname_rowname" where the name prior to the underscore is the column name and the name after the underscore is the row name.

ID
A variable name, as a character string, to use as names for the list elements.

effect_size
A vector of observed effect sizes to be modeled with metafor. See rma.mv for more details.

var_cor
Variance covariance matrix of sampling errors or sampling weights passed on to metafor. See rma.mv for more details.

weights
Optional weights matrix passed to metafor. Default is NULL. See rma.mv for more details.

moderators
Moderators (covariates, predictors) to be included in the analysis with metafor. Default is NULL meaning no predictors. See rma.mv for more details.

random_params
Specification of the random effect structure. This argument takes the form a single sided formula. Default is NULL which means a fixed effects model will be fitted with metafor. See rma.mv for more details.

structure
The variance structure passed to metafor. Default is "CS" which represents compound symmetry. See rma.mv for more details.

test
What type of test statistic should be used. Default is a t-statistic. See rma.mv for more details.

intercept
Should an intercept be included in the metafor model. See rma.mv for more details.

estimation_method
Estimation method to pass on to metafor. Default is "REML". See rma.mv for more details.

model
This is model syntax specified in the format by lavaan.

num_obs
Number of observations

...
Additional parameters passed to metafor. See rma.mv for more details.

See also